+++ packages/app-store/routing-forms/lib/handleResponse.ts
@@ -13,8 +13,8 @@ import type { ZResponseInputSchema } from "@calcom/trpc/server/routers/viewer/ro
 import { TRPCError } from "@trpc/server";
 
 import isRouter from "../lib/isRouter";
-import { onFormSubmission } from "../trpc/utils";
 import type { FormResponse, SerializableForm } from "../types/types";
+import { onFormSubmission } from "../utils";
 
 export type Form = SerializableForm<
   App_RoutingForms_Form & {
